refactor: businessprofile 후기 폴더구조 변경(#112)
