+++ backend/src/docs/asciidoc/post.adoc
@@ -188,17 +188,3 @@ include::{snippets}/post-apply-reject/path-parameters.adoc[]
 
 성공 응답 HTTP
 include::{snippets}/post-apply-reject/http-response.adoc[]
-
-== 게시물 후기 관련 API
-=== 게시물 후기 작성하기
-
-`POST /api/post/{postId}/review/{postDoDateId}`
-
-요청 HTTP
-include::{snippets}/post-review-create/http-request.adoc[]
-include::{snippets}/post-review-create/path-parameters.adoc[]
-include::{snippets}/post-review-create/request-fields.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/post-review-create/http-response.adoc[]
-include::{snippets}/post-review-create/response-fields.adoc[]
