+++ backend/src/main/java/com/minibeit/businessprofile/dto/BusinessProfilesReviewRequest.java
@@ -1,0 +1,19 @@
+package com.minibeit.businessprofile.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+public class BusinessProfilesReviewRequest {
+    @Getter
+    @Builder
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    public static class CreateReview {
+        private String postTitle;
+        private String content;
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm", timezone = "Asia/Seoul")
+        private LocalDateTime doDate;
+    }
+}
