+++ backend/src/main/java/com/minibeit/post/dto/PostRequest.java
@@ -43,15 +43,4 @@ public class PostRequest {
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm", timezone = "Asia/Seoul")
         private List<LocalDateTime> doDateList;
     }
-
-    @Getter
-    @Builder
-    @NoArgsConstructor(access = AccessLevel.PRIVATE)
-    @AllArgsConstructor(access = AccessLevel.PRIVATE)
-    public static class CreateReview {
-        private String postTitle;
-        private String content;
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm", timezone = "Asia/Seoul")
-        private LocalDateTime doDate;
-    }
 }
